[test_utils] move StubBrowserStorage to a more central location
