+++ src/test_utils/stub_browser_storage.js
