+++ src/ui/public/chrome/api/__tests__/apps.js
@@ -1,7 +1,7 @@
 import expect from 'expect.js';
 
 import setup from '../apps';
-import StubBrowserStorage from '../../__tests__/fixtures/stub_browser_storage';
+import StubBrowserStorage from 'test_utils/stub_browser_storage';
 
 describe('Chrome API :: apps', function () {
   describe('#get/setShowAppsLink()', function () {
